Documentacion con io.swagger añadida a catalogo

+++ catalogo/src/main/java/com/catalogo/application/resources/ActorResource.java
@@ -6,6 +6,7 @@ import java.util.List;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ProblemDetail;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -28,10 +29,18 @@ import com.catalogo.exceptions.DuplicateKeyException;
 import com.catalogo.exceptions.InvalidDataException;
 import com.catalogo.exceptions.NotFoundException;
 
+import io.swagger.v3.oas.annotations.Hidden;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 
 @RestController
+@Tag(name = "actores-service", description = "Mantenimiento de actores")
 @RequestMapping("/api/actores/v1")
 public class ActorResource {
 	private ActorService srv;
@@ -40,8 +49,12 @@ public class ActorResource {
 		this.srv = srv;
 	}
 	
+	@Operation(summary = "Listado de todos los actores")
+	@ApiResponse(responseCode = "200", description = "OK")
 	@GetMapping
-	public List getAll(@RequestParam(required = false, defaultValue = "largo") String modo) {
+	public List getAll(
+			 @Parameter(description = "Modo de vista de los actores", required = false, schema = @Schema(allowableValues = {"short", "largo"}, defaultValue = "largo")) 
+			@RequestParam(required = false, defaultValue = "largo") String modo) {
 		if("short".equals(modo))
 			return srv.getByProjection(ActorShort.class);
 		else
@@ -51,13 +64,21 @@ public class ActorResource {
 	/* 
 	 * Muestra por defecto 20 elementos en cada página
 	 */
+	@Hidden
 	@GetMapping(params = "page")
-	public Page<ActorShort> getAll(Pageable page){
+	public Page<ActorShort> getAll(
+			@Parameter(hidden = true) Pageable page){
 		return srv.getByProjection(page, ActorShort.class);
 	}
 	
+	@Operation(summary = "Recupera un actor", description = "Está disponible en versión corta con los datos imprescindibles del actor.")
+	@ApiResponse(responseCode = "200", description = "Actor encontrado", content = @Content(schema = @Schema(oneOf = {
+			ActorDTO.class })))
+	@ApiResponse(responseCode = "404", description = "Actor no encontrado", content = @Content(schema = @Schema(implementation = ProblemDetail.class)))
 	@GetMapping(path = "/{id}")
-	public ActorDTO getOne(@PathVariable int id) throws NotFoundException{
+	public ActorDTO getOne(
+			@Parameter(description = "Identificador del actor", required = true) 
+			@PathVariable int id) throws NotFoundException{
 		var item = srv.getOne(id);
 		if(item.isEmpty())
 			throw new NotFoundException();
@@ -67,9 +88,14 @@ public class ActorResource {
 	//Un record se utiliza para crear una clase inmutable --> usar cuando solo se va a usar una vez
 	record Peli(int id, String titulo) {}
 	
+	@Operation(summary = "Listado de las peliculas del actor")
+	@ApiResponse(responseCode = "200", description = "Actor encontrado")
+	@ApiResponse(responseCode = "404", description = "Actor no encontrado")
 	@GetMapping(path = "/{id}/pelis")
 	@Transactional
-	public List<Peli> getPelis(@PathVariable int id) throws NotFoundException{
+	public List<Peli> getPelis(
+			@Parameter(description = "Identificador del actor", required = true) 
+			@PathVariable int id) throws NotFoundException{
 		var item = srv.getOne(id);
 		if(item.isEmpty())
 			throw new NotFoundException();
@@ -78,7 +104,12 @@ public class ActorResource {
 				.toList();
 	}
 	
+	@Operation(summary = "Añadir un nuevo actor")
+	@ApiResponse(responseCode = "201", description = "Actor añadido")
+	@ApiResponse(responseCode = "404", description = "Actor no encontrado")
 	@PostMapping
+	@ResponseStatus(code = HttpStatus.CREATED)
+	@Transactional
 	public ResponseEntity<Object> create(@Valid @RequestBody Actor item) throws BadRequestException, DuplicateKeyException, InvalidDataException {
 		var newItem = srv.add(item);
 		URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}")
@@ -86,7 +117,12 @@ public class ActorResource {
 		return ResponseEntity.created(location).build();	
 	}
 	
+	@Operation(summary = "Añadir un nuevo actor - version corta")
+	@ApiResponse(responseCode = "201", description = "Actor añadido")
+	@ApiResponse(responseCode = "404", description = "Actor no encontrado")
 	@PostMapping("/corto")
+	@ResponseStatus(code = HttpStatus.CREATED)
+	@Transactional
 	public ResponseEntity<Object> createCorto(@Valid @RequestBody ActorDTO item) throws BadRequestException, DuplicateKeyException, InvalidDataException {
 		var newItem = srv.add(ActorDTO.from(item));
 		URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}")
@@ -94,17 +130,30 @@ public class ActorResource {
 		return ResponseEntity.created(location).build();	
 	}
 	
+	@Operation(summary = "Modificar un actor existente", description = "Los identificadores deben coincidir")
+	@ApiResponse(responseCode = "200", description = "Actor encontrado")
+	@ApiResponse(responseCode = "400", description = "Datos inválidos", content = @Content(schema = @Schema(implementation = ProblemDetail.class)))
+	@ApiResponse(responseCode = "404", description = "Actor no encontrado")
 	@PutMapping(path = "/{id}")
 	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void update(@PathVariable int id, @Valid @RequestBody ActorDTO item) throws BadRequestException, NotFoundException,  InvalidDataException {
+	public void update(
+			@Parameter(description = "Identificador del actor", required = true) 
+			@PathVariable int id, @Valid @RequestBody ActorDTO item) throws BadRequestException, NotFoundException,  InvalidDataException {
 		if(id != item.getActorId())
 			throw new BadRequestException("No coinciden los identificadores");
 		srv.modify(ActorDTO.from(item));
 	}
 	
+	@Operation(summary = "Borrar un actor existente")
+	@ApiResponse(responseCode = "204", description = "Actor borrado")
+	@ApiResponse(responseCode = "404", description = "Actor no encontrado")
 	@DeleteMapping(path = "/{id}")
 	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void delete(@PathVariable int id) throws BadRequestException, NotFoundException,  InvalidDataException {
+	public void delete(
+			@Parameter(description = "Identificador del actor", required = true) 
+			@PathVariable int id) throws BadRequestException, NotFoundException,  InvalidDataException {
+		if(srv.getOne(id).isEmpty())
+    		throw new NotFoundException();
 		srv.deleteById(id);
 	}
